# docker_microservice2/swagger.yml
basePath: /
definitions:
  CommonError:
    properties:
      message:
        description: |-
          Message of the error
          in: string
        type: string
        x-go-name: Message
      status:
        description: |-
          Status of the error
          in: int64
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: golang-master/controllers
  CommonSuccess:
    properties:
      message:
        description: |-
          Message of the error
          in: string
        type: string
        x-go-name: Message
      status:
        description: |-
          Status of the error
          in: int64
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: golang-master/controllers
  Companies:
    items:
      $ref: '#/definitions/Company'
    type: array
    x-go-package: golang-master/models
  Company:
    properties:
      id:
        description: |-
          Id of the company
          in: int64
        format: int64
        type: integer
        x-go-name: Id
      name:
        description: |-
          Name of the company
          in: string
        type: string
        x-go-name: Name
      status:
        description: |-
          Status of the company
          in: int64
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: golang-master/models
  GetCompanies:
    properties:
      data:
        $ref: '#/definitions/Companies'
      message:
        description: |-
          Message of the error
          in: string
        type: string
        x-go-name: Message
      status:
        description: |-
          Status of the error
          in: int64
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: golang-master/controllers
  GetCompany:
    properties:
      data:
        $ref: '#/definitions/Company'
      message:
        description: |-
          Message of the error
          in: string
        type: string
        x-go-name: Message
      status:
        description: |-
          Status of the error
          in: int64
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: golang-master/controllers
host: localhost:5000
info:
  version: 0.0.2
paths:
  /admin/company/:
    post:
      description: Create a new company
      operationId: addCompany
      parameters:
      - description: 'name: Authorization'
        in: header
        name: authorization
        required: true
        type: string
        x-go-name: Authorization
      responses:
        "200":
          description: CommonSuccess
          schema:
            $ref: '#/definitions/CommonSuccess'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      tags:
      - admin
  /admin/company/{id}:
    delete:
      description: Delete company
      operationId: deleteCompany
      parameters:
      - description: 'name: id'
        in: path
        name: id
        required: true
        type: string
        x-go-name: Id
      - description: 'name: Authorization'
        in: header
        name: authorization
        required: true
        type: string
        x-go-name: Authorization
      responses:
        "200":
          description: CommonSuccess
          schema:
            $ref: '#/definitions/CommonSuccess'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      tags:
      - admin
    put:
      consumes:
      - application/x-www-form-urlencoded
      description: Edit a company
      operationId: editCompany
      parameters:
      - description: 'name: id'
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: Id
      - description: 'name: Name'
        example: test
        in: formData
        name: name
        required: true
        type: string
        x-go-name: Name
      - description: 'name: Status'
        example: 1
        format: int64
        in: formData
        name: status
        required: true
        type: integer
        x-go-name: Status
      - description: 'name: Authorization'
        in: header
        name: authorization
        required: true
        type: string
        x-go-name: Authorization
      responses:
        "200":
          description: GetCompany
          schema:
            $ref: '#/definitions/GetCompany'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      tags:
      - admin
  /admin/company/list:
    get:
      description: Get companies list
      operationId: listCompany
      parameters:
      - description: 'name: Authorization'
        in: header
        name: authorization
        required: true
        type: string
        x-go-name: Authorization
      responses:
        "200":
          description: GetCompanies
          schema:
            $ref: '#/definitions/GetCompanies'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      tags:
      - admin
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  Authorization:
    description: Requests should pass an api_key header.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"